frontend expense service and basic error handling util
