+++ apps/mobile/app/utils/serviceUtils.ts
@@ -1,0 +1,6 @@
+export const handleError = (error: any): never => {
+  const errorMessage =
+    error.response?.data?.message ||
+    'An unexpected error occurred with expense services.';
+  throw new Error(errorMessage);
+};
